feat: create dashboard type definition file
